% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gto.R
\name{gto_generator}
\alias{gto_generator}
\title{Function to generate Gaussian-Type Orbital Functions}
\usage{
gto_generator(alpha)
}
\arguments{
\item{alpha}{a numeric}
}
\value{

}
\description{
This function takes in a constant "alpha" and returns a function that takes
in the radius and gives you the energy at that radius.
}
\examples{
TRUE
}
