% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sto_3g.R
\name{sto_3g_generator}
\alias{sto_3g_generator}
\title{Function to generate STO-3G type orbital functions}
\usage{
sto_3g_generator(basis, element)
}
\arguments{
\item{element}{}
}
\value{

}
\description{
#' Function to generate Slater-Type Orbital Functions
}
\details{
This function takes in the name of an element and a basiss et and generates a function that can return the energy at that radius.
}
\examples{
TRUE
}
