% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sto_5g.R
\name{sto_5g_generator}
\alias{sto_5g_generator}
\title{Function to generate STO-5G type orbital functions}
\usage{
sto_5g_generator(element)
}
\arguments{
\item{element}{The name or symbol of the element}
}
\value{

}
\description{
This function takes in the name of an element and a basis et and generates a function that can return the energy at that radius.
}
\examples{
TRUE
}
