% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sto.R
\name{sto_generator}
\alias{sto_generator}
\title{Function to generate Slater-Type Orbital Functions}
\usage{
sto_generator(zeta)
}
\arguments{
\item{zeta}{a numeric}
}
\value{

}
\description{
This function takes in a constant "zeta" and returns a function that takes
in the radius and gives you the energy at that radius.
}
\examples{
TRUE
}
