% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_basis_set.R
\name{load_basis_set}
\alias{load_basis_set}
\title{Download basis sets}
\usage{
load_basis_set(basis, element)
}
\arguments{
\item{basis}{The basis set to use}

\item{element}{The name or the symbol of the element}
}
\value{

}
\description{
This function takes a name of an element and returns its basis set.
}
\examples{
TRUE
}
